moved JustificationSet to separate file decluttering
